search: widen countermove table from i16 to i32 destination coords
Destinations were truncated to i16, so on the unbounded board a piece
past +-32767 could alias to the wrong countermove entry. Fixed at both
the write site and the two read/compare sites (ordering.rs, movegen.rs).

No SPRT: self-play never reaches i16's range (measured max coord ~103),
so no played game can exercise the changed path either way; verified
instead with a unit test using an out-of-range coordinate plus the full
existing suite (504 passed).

+++ src/search/movegen.rs
@@ -558,8 +558,8 @@ impl StagedMoveGen {
             };
             if entry.0 != 0
                 && entry.0 == m.piece.piece_type() as u8
-                && entry.1 == m.to.x as i16
-                && entry.2 == m.to.y as i16
+                && entry.1 == m.to.x as i32
+                && entry.2 == m.to.y as i32
             {
                 score += sort_countermove();
             }
@@ -1358,8 +1358,8 @@ mod tests {
         searcher.prev_move_stack[0] = (3, 9);
         searcher.countermoves[3][9] = (
             quiet.piece.piece_type() as u8,
-            quiet.to.x as i16,
-            quiet.to.y as i16,
+            quiet.to.x as i32,
+            quiet.to.y as i32,
         );
         let picker = StagedMoveGen::new(None, 1, 2, &searcher, &game);
         assert!(picker.score_quiet(&game, &searcher, &quiet) >= sort_countermove());
